Update Login System
Fixed the previous problem. Is clunkier and repeats more code than I'd prefer, but it works now. Allows in when criteria filled, doesn't if criteria is not. 
Still need to add write to file and then checking compared to that which will then allow write to file for the username also.