Update Login System
Added the first set of write and read from file. Now, when the user creates an account their username and password is written to file and when they try to log in it reads the file, checking if their credentials match those stored - if it does, it lets them in. Still need to add a check on whether a username is already taken and also currently despite being logged in, it continues to repeat so I need to look at that