Update Login System
Added the check for repeated usernames. That part of the function works however it just continues to the password part regardless currently. To fix this I think there will have to be a larger change because the repeat is already an issue, so I may add a new function to call when it needs to be repeated rather than attempting to do it within the other functions because currently that's not working correctly