Update Bouncing Ball Game
Added the ball. Currently bounces off wall but need to make it also bounce off the platform. Then I need to decide whether to have floating things to collect for points or just how many times it hits platform to be points, as well as make speed increase over time.