Update Bouncing Ball Game
Added a 'Game Over' display when player lets the ball hit the bottom border. Also, found a bug where after the first bounce, it no longer ended game if it hit the bottom. I fixed this by making it >= instead of == 700 which seems to have fixed the issue. Next, I want to add objects to bounce off higher up in the screen and maybe some coins (or similar) to collect