Update Bouncing Ball Game
Attempted to add a score write to file and a high score system in. The write to file works - if the player's score is higher than stored, that becomes the score stored. However, currently even if this happens the score written on the screen at the end is still the previous, which I need to sort. I would also like to add a feature where after a certain point the platform speed also increases because eventually the ball will be too fast for the platform and it would mean some are impossible to get. Also, I would like to make a starting GUI to click to start.