Update Bouncing Ball Game
The game is finished. There aren't set levels, but the game does restart after each fail and if you get bored you can change where the blocks are on the screen. I didn't feel changing the colour of the platforms was necessary. The graphics flicker a bit, but the game is very playable still besides that.